Don't try to install the emscripten SDK.
We already assume the user has a working compiler, so making the setup
of emscripten itself the user's responsibility makes sense. Plus there
are multiple ways it can be set up; Debian has a package for example.

This changes the emscripten pkgdef to just install wrapper scripts and
ditch the code to set up the SDK.